Update app/controllers/student_task_controller.rb
1. The list method has been shortened and three new methods have been made namely
a.	task_stage_submission:  Decides whether to display the tasks in submission stage as part of 'tasks not yet started' or 'tasks in revision'
b.	task_stage_review: Decides whether to display the tasks in review stage as part of 'tasks not yet started' or 'tasks in revision'
c.	 task_stage_metareview: Decides whether to display the tasks in metareview stage as part of 'tasks not yet started' or 'tasks in revision'
2. Also new methods  task_not_yet_started and tasks_in_revision have been added to remove the extra code from the view list.html.erb:
a. tasks_not_yet_started: This method controls the redirection on clicking on a task in the left navigation pane under tasks not yet started
b. tasks_in_revision: This method controls the redirection on clicking on a task in the left navigation pane under tasks in revision